Create Anvil.debugRender(long threshould)
This function allow to see which views are rendering too slow.
The anvil will automatically paints with red color bad views, that
executes the Renderable.view() method above a threshould in milliseconds.

The debug mode also logs the view mount time in the console in the
following format:

11-24 20:58:26.904 15181 15181 I System.out: BENCHMARCK ANVIL MainActionBar rendered in 1 ms
11-24 20:58:26.904 15181 15181 I System.out: BENCHMARCK ANVIL SomeBadView rendered in 100 ms

This option should be called in the Application.onCreate(...).